# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MTC Group Staff Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: assignee_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: department_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  # ─── Documents ───────────────────────────────────────────────────────────────
...
  /policy/current:
    get:
      operationId: getCurrentPolicy
      tags: [policy]
      summary: Get the current published policy version and acknowledgment status
      responses:
        "200":
          description: Current policy with acknowledgment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyWithAck"
        "404":
          description: No policy published yet
...
  /policy:
    post:
      operationId: publishPolicy
      tags: [policy]
      summary: Publish a new policy version (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishPolicyBody"
      responses:
        "201":
          description: Policy published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyVersion"

  # ─── Auth / Login Event ──────────────────────────────────────────────────────
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for the current user (inbox or sent)
      parameters:
        - name: box
          in: query
          schema:
            type: string
            enum: [inbox, sent]
      responses:
        "200":
          description: Message list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageItem"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a new internal message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageItem"
  /messages/{id}/read:
    patch:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark a message as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageItem"
  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message (soft delete for sender or recipient)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - assigneeId
        - assigneeName
        - departmentId
        - status
        - dueDate
        - createdAt
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
      required:
        - title
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        dueDate:
          type: ["string", "null"]
...
    PolicyVersion:
      type: object
      properties:
        id:
          type: integer
        version:
          type: string
        body:
          type: string
        publishedAt:
          type: string
        publishedBy:
          type: ["integer", "null"]
      required:
        - id
        - version
        - body
        - publishedAt
        - publishedBy
    PolicyWithAck:
      type: object
      properties:
        id:
          type: integer
        version:
          type: string
        body:
          type: string
        publishedAt:
          type: string
        publishedBy:
          type: ["integer", "null"]
        acknowledgedAt:
          type: ["string", "null"]
      required:
        - id
        - version
        - body
        - publishedAt
        - publishedBy
        - acknowledgedAt
    PublishPolicyBody:
      type: object
      properties:
        version:
          type: string
        body:
          type: string
      required:
        - version
        - body
...
    MessageItem:
      type: object
      properties:
        id:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: ["integer", "null"]
        fromName:
          type: string
        toName:
          type: string
        subject:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        scope:
          type: string
        createdAt:
          type: string
      required:
        - id
        - fromUserId
        - toUserId
        - fromName
        - toName
        - subject
        - body
        - isRead
        - scope
        - createdAt
    SendMessageBody:
      type: object
      properties:
        toUserId:
          type: ["integer", "null"]
        subject:
          type: string
        body:
          type: string
        scope:
          type: string
      required:
        - subject
        - body
...